fixed bugs

+++ src/models/users.model.js
@@ -328,11 +328,13 @@ const Users = SequelizeConnector.define(
         }
       },
       afterDestroy: async (user, { transaction }) => {
-        await Otps.destroy({
-          where: { phoneCountryCode: user.phoneCountryCode, phoneNumber: user.number },
-          transaction,
-          force: true
-        });
+        if (user.phoneCountryCode && user.phoneNumber) {
+          await Otps.destroy({
+            where: { phoneCountryCode: user.phoneCountryCode, phoneNumber: user.phoneNumber },
+            transaction,
+            force: true
+          });
+        }
       }
     }
   }
